--- src/main/java/store/OutputFormatter.java
@@ -1,7 +1,0 @@
-package store;
-
-public class OutputFormatter {
-    public static String formatWithNegativeSign(int price) {
-        return String.format("-%,d", Math.abs(price));
-    }
-}
