+++ src/main/java/store/inventory/Promotion.java
@@ -1,4 +1,4 @@
-package store;
+package store.inventory;
 
 import camp.nextstep.edu.missionutils.DateTimes;
 import java.time.LocalDate;
