+++ src/main/java/store/purchase/Membership.java
@@ -1,4 +1,4 @@
-package store;
+package store.purchase;
 
 public class Membership {
     private final int price;
