+++ src/main/java/store/view/OutputView.java
@@ -1,6 +1,10 @@
-package store;
+package store.view;
 
 import java.util.List;
+import store.inventory.Product;
+import store.inventory.PromotionType;
+import store.inventory.Stock;
+import store.purchase.Receipt;
 
 public class OutputView {
     private static final String START_PROMPT = "안녕하세요. W편의점입니다.";
