+++ src/test/java/store/common/AnswerValidatorTest.java
@@ -1,4 +1,4 @@
-package store;
+package store.common;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
