+++ src/test/java/store/inventory/StockTest.java
@@ -1,4 +1,4 @@
-package store;
+package store.inventory;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
