+++ src/test/java/store/purchase/PurchaseParserTest.java
@@ -1,4 +1,4 @@
-package store;
+package store.purchase;
 
 import java.util.Map;
 import org.assertj.core.api.Assertions;
